requirements.txt
--find-links https://download.pytorch.org/whl/nightly/cu128
torch>=2.2.0.dev0
torchvision>=0.17.0.dev0
torchaudio>=2.2.0.dev0
transformers>=4.36.0
sentence-transformers>=2.2.2
numpy>=1.24.0
scikit-learn>=1.3.0
tqdm>=4.65.0
colorama>=0.4.6
python-dotenv>=1.0.0
python-docx>=0.8.11
PyPDF2>=3.0.0
chardet>=5.0.0 